+++ CHANGES.md
@@ -1,4 +1,5 @@
 #### [unreleased]
+* only use `esc_attr_e` for translating strings
 
 #### 10.4.2 / 2021-07-21
 * always get current release asset redirect URL on REST update