+++ CHANGES.md
@@ -1,5 +1,6 @@
 #### [unreleased]
 * only use `esc_attr_e` for translating strings
+* ksort additions into plugin/theme array
 
 #### 10.4.2 / 2021-07-21
 * always get current release asset redirect URL on REST update