+++ CHANGES.md
@@ -3,6 +3,7 @@
 * ksort additions into plugin/theme array
 * use `gu_config_pre_process` filter in `update_site_transient()`
 * remove vanity star ratings
+* speed up `get_dot_org_data()` by using API 1.2 and `wp_remote_head()`
 
 #### 10.4.2 / 2021-07-21
 * always get current release asset redirect URL on REST update